use grpc.WithRPCCredentials rather than client interceptor
