+++ pkg/auth/grpcauth/apikey_test.go
@@ -1,0 +1,15 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+package grpcauth
+
+import (
+	"testing"
+)
+
+func TestAPIKey(t *testing.T) {
+	// todo implement test
+	// {"", status.Errorf(codes.Unauthenticated, "Invalid API credential")}
+	// {"good key", nil}
+	// {"wrong key", status.Errorf(codes.Unauthenticated, "Invalid API credential")}
+}
